+++ src/niles/lwjgl/entity/Entity.java
@@ -36,8 +36,18 @@ public class Entity {
 		textures.add(texture);
 	}
 	public void bindTextures() {
+		if(textures.size() == 0) {
+			Texture.unbind();
+		}
 		for(int i = 0; i < textures.size(); i++) {
 			textures.get(i).bind(i);
+		}
+	}
+	
+	//textures should be deleted when not used anymore.
+	public void DeleteTextures() {
+		for(int i = 0; i < textures.size(); i++) {
+			textures.get(i).delete();
 		}
 	}