+++ src/niles/lwjgl/loop/Game.java
@@ -7,17 +7,11 @@ import static org.lwjgl.glfw.GLFW.GLFW_KEY_Q;
 import static org.lwjgl.glfw.GLFW.GLFW_KEY_S;
 import static org.lwjgl.glfw.GLFW.GLFW_KEY_W;
 
-import java.util.HashMap;
+import java.util.ArrayList;
 
 import org.joml.Vector3f;
 import org.joml.Vector4f;
 
-import niles.lwjgl.entity.Entity;
-import niles.lwjgl.fbo.Fbo;
-import niles.lwjgl.light.Lights;
-import niles.lwjgl.npsl.Shader;
-import niles.lwjgl.rendering.Renderer;
-import niles.lwjgl.world.Camera;
 import niles.lwjgl.world.Input;
 import niles.lwjgl.world.Mouse;
 import niles.lwjgl.world.Window;
@@ -25,36 +19,19 @@ import niles.lwjgl.world.Window;
 public abstract class Game {
 	
 	private Window window;
-	private Camera camera;
-	private Renderer renderer;
 	private Input input;
-	
-	private Lights lights;
-	private Fbo fbo;
 	
 	private Vector4f backgroundColor;
 	private int fpsCap;
 	
-	public Game(int width,int height,boolean fullsceen,Vector4f backgroundColor,int fpsCap) {
-		window=new Window(width, height, fullsceen);
-		camera=new Camera();
-		input = new Input(getWindow());
-		lights = new Lights();
-		fbo = new Fbo();
-		
-		this.backgroundColor=backgroundColor;
-		this.fpsCap=fpsCap;
-		
-		loop();
-	}
+	private ArrayList<Scene> scenes;
+	private int currentScene = 0;
 	
-	public Game() {
-		window=new Window(1920, 1080, true);
-		camera=new Camera();
-		camera.setPerspective((float) Math.toRadians(70), 1920f / 1080f, 0.1f, 1000);
+	
+	public Game(int width,int height,boolean fullsceen) {
+		window=new Window(width, height, fullsceen);
 		input = new Input(getWindow());
-		lights = new Lights();
-		fbo = new Fbo();
+		scenes = new ArrayList<Scene>();
 		
 		this.backgroundColor=new Vector4f(0,0,0,1);
 		this.fpsCap=120;
@@ -62,120 +39,80 @@ public abstract class Game {
 		loop();
 	}
 	
-	public abstract void setup();
-	
-	public abstract void update();
-	
-	public void loop() {
-		renderer = new Renderer();
+	public Game() {
+		window=new Window(1920, 1080, true);
+		input = new Input(getWindow());
 		
-		setup();
+		scenes = new ArrayList<Scene>();
+		
+		this.backgroundColor=new Vector4f(0,0,0,1);
+		this.fpsCap=120;
+		
+		loop();
+	}
+	
+	public abstract void init();
+	
+	private void loop() {
+		
+		init();
 		
 		while(window.shouldUpdate()) {
 			window.drawInit(backgroundColor);
 			
-			update();
+			if(scenes.size() > 0) {
+				scenes.get(currentScene).loop();
+			}
 			
-			renderer.clean();
 			window.clean();
 			window.update(fpsCap);
 		}
 	}
 	
-	public void render(Entity entity) {
-		renderer.render(getCamera(), entity, lights);
-	}
-	
-	public void rotateCamera(float xAxis, float yAxis) {
-		getCamera().getRotation().setAngleAxis(yAxis, 0, 1, 0);
-		getCamera().getRotation().rotate(xAxis, 0, 0);
-	}
-	
-	public void moveCameraForward(float amount) {
-		Vector3f move = new Vector3f(0, 0, -amount);
-		move.rotate(getCamera().getRotation());
-		getCamera().getPosition().add(move);
-	}
-	public void moveCameraBackward(float amount) {
-		Vector3f move = new Vector3f(0, 0, amount);
-		move.rotate(getCamera().getRotation());
-		getCamera().getPosition().add(move);
-	}
-	public void moveCameraLeft(float amount) {
-		Vector3f move = new Vector3f(-amount, 0, 0);
-		move.rotate(getCamera().getRotation());
-		getCamera().getPosition().add(move);
-	}
-	public void moveCameraRight(float amount) {
-		Vector3f move = new Vector3f(amount, 0, 0);
-		move.rotate(getCamera().getRotation());
-		getCamera().getPosition().add(move);
-	}
-	public void moveCameraUp(float amount) {
-		Vector3f move = new Vector3f(0, amount, 0);
-		move.rotate(getCamera().getRotation());
-		getCamera().getPosition().add(move);
-	}
-	public void moveCameraDown(float amount) {
-		Vector3f move = new Vector3f(0, -amount, 0);
-		move.rotate(getCamera().getRotation());
-		getCamera().getPosition().add(move);
-	}
-	
 	
 	public void simpleCameraRotation(float sensitivity) {
 		Mouse.isVisible(getWindow(), false);
-		Mouse.moveMouse(getWindow(), 1f);
-		rotateCamera(-Mouse.myY, -Mouse.myX);
+		Mouse.moveMouse(getWindow(), sensitivity);
+		scenes.get(currentScene).rotateCamera(-Mouse.y, -Mouse.x);
 	}
 	
 	public void simpleCameraMovement(float speed) {
 		if(getInput().isDown(GLFW_KEY_W)) {
-			moveCameraForward(speed);
+			scenes.get(currentScene).moveCameraForward(speed);
 		}
 		if(getInput().isDown(GLFW_KEY_S)) {
-			moveCameraBackward(speed);
+			scenes.get(currentScene).moveCameraBackward(speed);
 		}
 		if(getInput().isDown(GLFW_KEY_A)) {
-			moveCameraLeft(speed);
+			scenes.get(currentScene).moveCameraLeft(speed);
 		}
 		if(getInput().isDown(GLFW_KEY_D)) {
-			moveCameraRight(speed);
+			scenes.get(currentScene).moveCameraRight(speed);
 		}
 		if(getInput().isDown(GLFW_KEY_Q)) {
-			moveCameraDown(speed);
+			scenes.get(currentScene).moveCameraDown(speed);
 		}
 		if(getInput().isDown(GLFW_KEY_E)) {
-			moveCameraUp(speed);
+			scenes.get(currentScene).moveCameraUp(speed);
 		}
 	}
-
 	
-	//fbo is used for post processing
-	public void renderFbo(Shader shader) {
-		fbo.render(shader, getCamera(), getLights());
+	
+	public void addScene(Scene scene) {
+		scenes.add(scene);
 	}
 	
-	public void bindFbo() {
-		fbo.bind();
-	}
-	
-	public void unbindFbo() {
-		fbo.unbind();
-	}
-	
-	
-	public void setFboUniform(String name, Object value) {
-		fbo.setUniform(name, value);
+	public void useScene(int sceneIndex) {
+		Mouse.x = 0;
+		Mouse.y = 0;
+		scenes.get(currentScene).clean();
+		
+		currentScene = sceneIndex;
 	}
 	
 	
 	public Window getWindow() {
 		return window;
-	}
-
-	public Camera getCamera() {
-		return camera;
 	}
 
 	public Vector4f getBackgroundColor() {
@@ -198,25 +135,9 @@ public abstract class Game {
 	public int getFps() {
 		return window.getFps();
 	}
-
-	public Renderer getRenderer() {
-		return renderer;
-	}
-
 	public Input getInput() {
 		return input;
 	}
 
-	public Lights getLights() {
-		return lights;
-	}
-
-	public void setLights(Lights lights) {
-		this.lights = lights;
-	}
-
-
-	
-	
 
 }